backend/requirements.txt
Django==4.2.10
django-cors-headers==4.3.1
graphene-django==3.1.5
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
python-dotenv==1.0.0
graphene==3.3
oauthlib==3.2.2
requests-oauthlib==1.3.1
PyJWT==2.8.0 